Fix warnings and small bug

+++ StatePathfinder.cpp
@@ -224,7 +224,7 @@ void StatePathfinder::clearPathSearch() {
 }
 
 void StatePathfinder::clearPathLine() {
-	for (int i = 0; i < path.size(); i++) {
+	for (unsigned i = 0; i < path.size(); i++) {
 		for (auto &node : path[i])
 			node->GetMesh()->RemoveModel(node);
 
@@ -268,18 +268,18 @@ void StatePathfinder::displayPathBezier(std::vector<Vec2<>> p, string id) {
 		
 		int steps = 30;
 
-		for (int i = 0; i < p.size(); i++) {
+		for (unsigned i = 0; i < p.size(); i++) {
 			Vec2<> ps = (i > 0) ? p[i - 1] : p.front();
 			Vec2<> pm = p[i];
 			Vec2<> pf = (i < p.size() - 1) ? p[i + 1] : p.back();
 
-			Vec2<float> dir_a = Vec2<float>(pm.x, pm.y) - Vec2<float>(ps.x, ps.y);
-			Vec2<float> dir_b = Vec2<float>(pf.x, pf.y) - Vec2<float>(pm.x, pm.y);
+			Vec2<float> dir_a = Vec2<float>((float)pm.x, (float)pm.y) - Vec2<float>((float)ps.x, (float)ps.y);
+			Vec2<float> dir_b = Vec2<float>((float)pf.x, (float)pf.y) - Vec2<float>((float)pm.x, (float)pm.y);
 
 			// Corner ahead
 			if (ps.x != pf.x && ps.y != pf.y) {
-				Vec2<float> p1 = Vec2<float>(ps.x, ps.y) + dir_a / 2.0f;
-				Vec2<float> p2(pm.x, pm.y);
+				Vec2<float> p1 = Vec2<float>((float)ps.x, (float)ps.y) + dir_a / 2.0f;
+				Vec2<float> p2((float)pm.x, (float)pm.y);
 				Vec2<float> p3 = p2 + dir_b / 2.0f;
 				Vec2<float> p4 = p3;
 
@@ -311,15 +311,15 @@ void StatePathfinder::displayPathCatmullRom(std::vector<Vec2<>> p, string id) {
 
 		int steps = 14;
 
-		for (int i = 0; i < p.size(); i++) {
+		for (unsigned i = 0; i < p.size(); i++) {
 			Vec2<> p1 = (i > 0) ? p[i - 1] : p[0];
 			Vec2<> p2 = p[i];
 			Vec2<> p3 = (i < p.size() - 1) ? p[i + 1] : p.back();
 			Vec2<> p4 = (i < p.size() - 2) ? p[i + 2] : p.back();
 
 			for (int i = 0; i < steps; i++) {
-				float cx = cspline(p1.x, p2.x, p3.x, p4.x, (float)i / (float)steps);
-				float cy = cspline(p1.y, p2.y, p3.y, p4.y, (float)i / (float)steps);
+				float cx = cspline((float)p1.x, (float)p2.x, (float)p3.x, (float)p4.x, (float)i / (float)steps);
+				float cy = cspline((float)p1.y, (float)p2.y, (float)p3.y, (float)p4.y, (float)i / (float)steps);
 
 				path[pathNum - 1].push_back(meshes[id]->CreateModel(cx * scale + origin.x, origin.y + 5.0f + (pathNum), cy * scale + origin.z));
 				path[pathNum - 1].back()->SetSkin(settings.getModels()[id].tex);
